worker/requirements_worker.txt
fastapi
uvicorn[standard]
opencv-python-headless
insightface
onnxruntime-gpu
ultralytics
python-dotenv
httpx
numpy
pillow
python-multipart
